Build example app using a single IIFE (#37)
Something isn’t working with Rollup and SystemJS. After a few hours I can’t figure out what the problem is and I’m not sure it’s worth spending more time.

This Rollup config is simpler and works. It’s not quite as optimal because all the dependencies are bundled in, and it must rebuild them on every change.